precision_labeling/pyproject.toml
[project]
name = "precision-labeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "emlearn>=0.23.1",
    "lightning>=2.6.0",
    "m2cgen>=0.10.0",
    "mlflow>=3.7.0",
    "nanobind>=2.10.2",
    "onnx>=1.20.0",
    "onnxscript>=0.5.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "polars>=1.36.1",
    "pyqt6>=6.10.1",
    "scikit-learn>=1.8.0",
    "tl2cgen>=1.0.0",
    "torchvision>=0.24.1",
    "xgboost>=3.1.2",
]
